Copy from PQClean commit b380047d7539a736a968b01a2e8a267955d18b7d

+++ src/kem/newhope/pqclean_newhope1024cca_clean/verify.c
@@ -22,7 +22,7 @@ int PQCLEAN_NEWHOPE1024CCA_CLEAN_verify(const unsigned char *a, const unsigned c
         r |= a[i] ^ b[i];
     }
 
-    r = (-(int64_t)r) >> 63;
+    r = (uint64_t)(-(int64_t)r) >> 63;
     return (int)r;
 }
